Revert "add user model,points and their tests"
