+++ api/__init__.py
@@ -1,6 +1,4 @@
 from flask import Flask, redirect
-
-from api.v1.users.views import blue_print_users
 from .v1.rides.views import blue_print_rides
 from .v1.requests.ride_requests.views import blue_print_ride_requests
 from .v1.notifications.views import blue_print_notifications
@@ -29,4 +27,3 @@ def index():
 app.register_blueprint(blue_print_rides)
 app.register_blueprint(blue_print_ride_requests)
 app.register_blueprint(blue_print_notifications)
-app.register_blueprint(blue_print_users)
