--- api/tests/__init__.py
@@ -1,4 +1,0 @@
-import unittest
-
-if __name__ == '__main':
-    unittest.main()
